requirements.txt
# Core Flask Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Database
Flask-PyMongo==2.3.0
pymongo==4.5.0

# Authentication & Security
authlib==1.6.0
httpx==0.28.1

# Image Processing & Storage
boto3==1.39.3
Pillow==11.3.0

# Real-time Features
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3

# Rate Limiting (Optional - currently disabled)
Flask-Limiter==3.5.0
redis==5.0.1

# Configuration & Environment
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Production Server
gunicorn==21.2.0